Migration: restaurant ratings are now decimals + a few styling tweaks

+++ src/server/models/restaurant.js
@@ -2,7 +2,7 @@
 module.exports = (sequelize, DataTypes) => {
   const restaurant = sequelize.define('restaurant', {
     name: DataTypes.STRING,
-    rating: DataTypes.INTEGER,
+    rating: DataTypes.DECIMAL,
     phone: DataTypes.STRING,
     address: DataTypes.STRING,
     tags: DataTypes.ARRAY(DataTypes.STRING),
